Projet DevOps/P_Kubernetes/Application/app/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.27
requests==2.26.0
Werkzeug==2.0.2
